@@ -14,7 +14,7 @@ description: Quran Search Engine and Statistics.
 version: 0.0.0+0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
