@@ -1,8 +1,8 @@
 import '../helpers/localization.dart';
 
 class ChapterSimpleDTO{
-  final int chapterID;
-  final String chapterNameAR;
+  final int? chapterID;
+  final String? chapterNameAR;
 
   ChapterSimpleDTO(this.chapterID, this.chapterNameAR);
 
