@@ -2,7 +2,7 @@ import '../dtos/verse-dto.dart';
 
 class VerseCollection{
   final List<VerseDTO> verses;
-  final String collectionName;
+  final String? collectionName;
 
   VerseCollection(this.verses, this.collectionName);
 }
