@@ -1,9 +1,9 @@
 class VerseDTO {
-  final String chapterName;
-  final String verseText;
-  final String verseTextTashkel;
-  final int verseID;
-  final int chapterId;
+  final String? chapterName;
+  final String? verseText;
+  final String? verseTextTashkel;
+  final int? verseID;
+  final int? chapterId;
   bool isSelected = false;
   bool isBookmark = false;
 
