@@ -2,7 +2,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 class CustomSharedPreferences {
   static CustomSharedPreferences instance = CustomSharedPreferences._();
-  SharedPreferences _preferences;
+  late SharedPreferences _preferences;
   static const String _BOOKMARK_CHAPTER_ID_KEY =
       "yatadabaron_bookmark_chapter_id";
   static const String _BOOKMARK_VERSE_ID_KEY = "yatadabaron_bookmark_verse_id";
@@ -16,16 +16,16 @@ class CustomSharedPreferences {
   }
 
   //Helpers
-  Future<String> _getString(String k) async {
-    try {
-      return this._preferences.getString(k);
-    } catch (e) {
-      print(e);
-      return null;
-    }
-  }
+  // Future<String?> _getString(String k) async {
+  //   try {
+  //     return this._preferences.getString(k);
+  //   } catch (e) {
+  //     print(e);
+  //     return null;
+  //   }
+  // }
 
-  Future<int> _getInt(String k) async {
+  Future<int?> _getInt(String k) async {
     try {
       return this._preferences.getInt(k);
     } catch (e) {
@@ -35,7 +35,7 @@ class CustomSharedPreferences {
   }
 
   //Get the saved bookmark chapter
-  Future<int> getBookmarkChapter() async {
+  Future<int?> getBookmarkChapter() async {
     return this._getInt(_BOOKMARK_CHAPTER_ID_KEY);
   }
 
@@ -44,7 +44,7 @@ class CustomSharedPreferences {
   }
 
   //Get the saved bookmark verse
-  Future<int> getBookmarkVerse() async {
+  Future<int?> getBookmarkVerse() async {
     return this._getInt(_BOOKMARK_VERSE_ID_KEY);
   }
 
