@@ -6,7 +6,7 @@ class ArabicNumbersService {
 
   ArabicNumbers _converter = ArabicNumbers();
 
-  String convert(int x, {bool reverse = true}) {
+  String convert(int? x, {bool reverse = true}) {
     String converted = _converter.convert(x);
     if (reverse == true) {
       String reversed = converted.split('').reversed.join();
