@@ -1,11 +1,9 @@
-import 'package:Yatadabaron/dtos/verse-dto.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../../blocs/search-session-bloc.dart';
 import '../../dtos/search-session-payload.dart';
 import '../../views/shared-widgets/loading-widget.dart';
-import 'package:share/share.dart';
 
 class SearchSummaryWidget extends StatelessWidget {
   @override
@@ -15,7 +13,7 @@ class SearchSummaryWidget extends StatelessWidget {
       stream: bloc.payloadStream,
       builder:
           (BuildContext context, AsyncSnapshot<SearchSessionPayload> snapshot) {
-        SearchSessionPayload searchSummary = snapshot.data;
+        SearchSessionPayload? searchSummary = snapshot.data;
         if (searchSummary == null) {
           return LoadingWidget();
         }
@@ -41,7 +39,7 @@ class SearchSummaryWidget extends StatelessWidget {
                 child: FloatingActionButton(
                   child: Icon(Icons.share),
                   onPressed: () async {
-                    snapshot.data.copyAll();
+                    snapshot.data!.copyAll();
                   },
                   mini: true,
                   heroTag: null,
