@@ -21,7 +21,7 @@ class ChaptersDropDown extends StatelessWidget {
         if (!snapshot.hasData) {
           return LoadingWidget();
         }
-        List<ChapterSimpleDTO> chapters = snapshot.data;
+        List<ChapterSimpleDTO> chapters = snapshot.data!;
         return Container(
           height: 300,
           width: double.maxFinite,
@@ -34,7 +34,7 @@ class ChaptersDropDown extends StatelessWidget {
                   .convert(chapter.chapterID, reverse: false);
               return ListTile(
                 title: Text(
-                  chapter.chapterNameAR,
+                  chapter.chapterNameAR!,
                   style: TextStyle(
                     fontSize: 18,
                   ),
